# src/AterStudio/ClientApp/src/app/share/models/project/template-file.model.ts
import { Project } from '../project/project.model';
/**
 * 模板内容
 */
export interface TemplateFile {
  id: string;
  /**
   * 名称
   */
  name: string;
  /**
   * 显示名称
   */
  displayName?: string | null;
  /**
   * 内容
   */
  content?: string | null;
  /**
   * 项目
   */
  project?: Project | null;
  projectId: string;

}