# src/AterStudio/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ater Studio",
    "description": "API 文档",
    "version": "v1"
  },
  "paths": {
...
    "/api/ApiDoc/all/{id}": {
      "get": {
        "tags": [
          "ApiDoc"
        ],
        "summary": "获取项目文档",
        "operationId": "ApiDoc_List",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "项目id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiDocInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiDocInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiDocInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ApiDoc/{id}": {
      "get": {
        "tags": [
          "ApiDoc"
        ],
        "summary": "获取某个文档信息",
        "operationId": "ApiDoc_GetApiDocContent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocContent"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocContent"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocContent"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ApiDoc"
        ],
        "summary": "更新",
        "operationId": "ApiDoc_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiDocInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiDocInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApiDocInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocInfo"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ApiDoc"
        ],
        "summary": "删除",
        "operationId": "ApiDoc_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/ApiDoc/export/{id}": {
      "get": {
        "tags": [
          "ApiDoc"
        ],
        "summary": "导出markdown文档",
        "operationId": "ApiDoc_Export",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/ApiDoc": {
      "post": {
        "tags": [
          "ApiDoc"
        ],
        "summary": "添加",
        "operationId": "ApiDoc_Add",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiDocInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiDocInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApiDocInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDocInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/ApiDoc/component": {
      "post": {
        "tags": [
          "ApiDoc"
        ],
        "summary": "生成页面组件",
        "operationId": "ApiDoc_CreateUIComponent",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUIComponentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUIComponentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUIComponentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NgComponentInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NgComponentInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NgComponentInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiDocContent": {
        "type": "object",
        "properties": {
          "restApiGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RestApiGroup"
            },
            "description": "接口信息"
          },
          "modelInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityInfo"
            },
            "description": "所有请求及返回类型信息"
          },
          "openApiTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiDocTag"
            },
            "description": "tag信息"
          }
        },
        "additionalProperties": false,
        "description": "接口返回模型"
      },
      "ApiDocInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "文档名称"
          },
          "description": {
            "maxLength": 1000,
            "type": "string",
            "description": "文档描述",
            "nullable": true
          },
          "path": {
            "maxLength": 300,
            "type": "string",
            "description": "文档地址"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "接口文档"
      },
      "ApiDocTag": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ComponentType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Form",
          "Table"
        ],
        "x-enumData": [ ]
      },
      "ConfigData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "maxLength": 100,
            "type": "string",
            "description": "键"
          },
          "value": {
            "maxLength": 2000,
            "type": "string",
            "description": "值"
          },
          "valueType": {
            "$ref": "#/components/schemas/ValueType"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "配置"
      },
...
      "CreateUIComponentDto": {
        "type": "object",
        "properties": {
          "uiType": {
            "$ref": "#/components/schemas/UIType"
          },
          "componentType": {
            "$ref": "#/components/schemas/ComponentType"
          },
          "modelInfo": {
            "$ref": "#/components/schemas/EntityInfo"
          },
          "serviceName": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "生成组件模型"
      },
...
      "EntityInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "类名"
          },
          "namespaceName": {
            "maxLength": 100,
            "type": "string",
            "description": "命名空间",
            "nullable": true
          },
          "assemblyName": {
            "maxLength": 100,
            "type": "string",
            "description": "程序集名称",
            "nullable": true
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "类注释",
            "nullable": true
          },
          "summary": {
            "maxLength": 100,
            "type": "string",
            "description": "类注释",
            "nullable": true
          },
          "keyType": {
            "$ref": "#/components/schemas/EntityKeyType"
          },
          "isEnum": {
            "type": "boolean",
            "description": "是否为枚举类",
            "nullable": true
          },
          "isList": {
            "type": "boolean"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "propertyInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyInfo"
            },
            "description": "属性"
          }
        },
        "additionalProperties": false,
        "description": "defined entity model"
      },
      "EntityKeyType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Guid",
          "Int",
          "String"
        ],
        "x-enumData": [ ]
      },
...
      "NgComponentInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "routing": {
            "type": "string",
            "nullable": true
          },
          "tsContent": {
            "type": "string",
            "nullable": true
          },
          "htmlContent": {
            "type": "string",
            "nullable": true
          },
          "cssContent": {
            "type": "string",
            "nullable": true
          },
          "moduleName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OperationType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Get",
          "Put",
          "Post",
          "Delete",
          "Options",
          "Head",
          "Patch",
          "Trace"
        ],
        "x-enumData": [ ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "项目名称"
          },
          "displayName": {
            "maxLength": 100,
            "type": "string",
            "description": "显示名"
          },
          "path": {
            "maxLength": 200,
            "type": "string",
            "description": "路径"
          },
          "version": {
            "maxLength": 20,
            "type": "string",
            "description": "版本",
            "nullable": true
          },
          "solutionType": {
            "$ref": "#/components/schemas/SolutionType"
          },
          "entityInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityInfo"
            }
          },
          "apiDocInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiDocInfo"
            }
          },
          "configDatas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigData"
            }
          },
          "templateFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateFile"
            }
          }
        },
        "additionalProperties": false,
        "description": "项目"
      },
...
      "PropertyInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "isList": {
            "type": "boolean",
            "description": "是否是数组"
          },
          "isPublic": {
            "type": "boolean"
          },
          "isNavigation": {
            "type": "boolean",
            "description": "是否为导航属性"
          },
          "isJsonIgnore": {
            "type": "boolean"
          },
          "navigationName": {
            "type": "string",
            "description": "导航属性类名称",
            "nullable": true
          },
          "isComplexType": {
            "type": "boolean"
          },
          "hasMany": {
            "type": "boolean",
            "description": "导航属性的对应关系",
            "nullable": true
          },
          "isEnum": {
            "type": "boolean"
          },
          "hasSet": {
            "type": "boolean",
            "description": "是否包括set方法"
          },
          "attributeText": {
            "type": "string",
            "nullable": true
          },
          "commentXml": {
            "type": "string",
            "description": "xml comment",
            "nullable": true
          },
          "commentSummary": {
            "type": "string",
            "description": "comment summary",
            "nullable": true
          },
          "isRequired": {
            "type": "boolean",
            "description": "是否必须"
          },
          "isNullable": {
            "type": "boolean",
            "description": "可空?"
          },
          "minLength": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxLength": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isDecimal": {
            "type": "boolean"
          },
          "suffixContent": {
            "type": "string",
            "description": "尾缀，如#endregion",
            "nullable": true
          },
          "defaultValue": {
            "type": "string",
            "description": "默认值"
          },
          "entityInfo": {
            "$ref": "#/components/schemas/EntityInfo"
          },
          "entityInfoId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "RestApiGroup": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "apiInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RestApiInfo"
            }
          }
        },
        "additionalProperties": false,
        "description": "接口分组信息"
      },
      "RestApiInfo": {
        "type": "object",
        "properties": {
          "operationType": {
            "$ref": "#/components/schemas/OperationType"
          },
          "router": {
            "type": "string",
            "description": "路由"
          },
          "summary": {
            "type": "string",
            "description": "说明",
            "nullable": true
          },
          "tag": {
            "type": "string",
            "nullable": true
          },
          "operationId": {
            "type": "string"
          },
          "queryParameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyInfo"
            },
            "description": "请求查询参数",
            "nullable": true
          },
          "requestInfo": {
            "$ref": "#/components/schemas/EntityInfo"
          },
          "responseInfo": {
            "$ref": "#/components/schemas/EntityInfo"
          }
        },
        "additionalProperties": false,
        "description": "接口信息"
      },
      "SolutionType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "DotNet",
          "Node"
        ],
        "x-enumData": [ ]
      },
...
      "TemplateFile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "名称"
          },
          "displayName": {
            "maxLength": 60,
            "type": "string",
            "description": "显示名称",
            "nullable": true
          },
          "content": {
            "maxLength": 10000,
            "type": "string",
            "description": "内容",
            "nullable": true
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "模板内容"
      },
...
      "UIType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "description": "使用的UI组件库",
        "format": "int32",
        "x-enumNames": [
          "AngularMaterial",
          "NGZORRO",
          "VueElement"
        ],
        "x-enumData": [ ]
      },
...
      "ValueType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Integer",
          "Double",
          "String",
          "Boolean"
        ],
        "x-enumData": [ ]
      }
    }
  },
...
}